requirements.txt
altair==6.0.0
attrs==25.4.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
cloudpickle==3.1.2
contourpy==1.3.3
cycler==0.12.1
fonttools==4.62.0
gitdb==4.0.12
GitPython==3.1.46
-e git+https://github.com/reetting/HEARTTfailureriskG5@b097b2cd8ba7b5666d7ea8cf991f9e9db7959496#egg=heart_failure
idna==3.11
imbalanced-learn==0.14.1
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
lightgbm==4.6.0
llvmlite==0.46.0
MarkupSafe==3.0.3
matplotlib==3.10.8
narwhals==2.17.0
numba==0.64.0
numpy==2.4.3
nvidia-nccl-cu12==2.29.7
packaging==26.0
pandas==2.3.3
pillow==12.1.1
pluggy==1.6.0
protobuf==6.33.5
pyarrow==23.0.1
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
shap==0.51.0
six==1.17.0
sklearn-compat==0.1.5
slicer==0.0.8
smmap==5.0.3
streamlit==1.55.0
tenacity==9.1.4
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.4
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
xgboost==3.2.0
